Limit usage of start/end args in PDF cmd to slash cmd only
This is because traditional cmd is inflexible (to my knowledge) about accepting args without the use of parameter flags (can't because hybrid cmd). And so because URL arg is optional but less so than start/end, it must come first. That means that if the user uploads the file instead leaving the URL blank, the user may enter a start/end and incorrectly be parsed as a URL and start argument which wouldn't happen if the usage was done through slash cmds.