Handle descending selection ranges in pdf cmd
Descending selection ranges (e.g., "3-2") will result in no output from the parser for that range specifically. If that range is the only one inputted, this will result in a final selection range list of length 0. So check for that and if so, send error msg. Else, other valid selection ranges will be the only one considered while the invalid one is ignored without errors, which is fine.