Added confirm. prompt when using chatgpt cmd for simple translate
The main goal here is to create some friction and disincentivize Google Translate-style usage of ChatGPT as it is unnecessary costs. If the member were to instead prompt a translation request but with explanation, that's considered a valid use-case for ChatGPT.

FIXME: The logic for the confirmation prompt (this commit) only works on non-slash usage. Figure out if possible. Alternatively, do we really need a slash version of this cmd?