Set `PYTHONUNBUFFERED` environment variable in docker-compose.yml
Python buffers its output and so the print stdout do not appear in Docker's logs immediately (although will eventually print if the buffer grows large enough and flushes). Fixed by setting PYTHONUNBUFFERED environment variable to a non-empty string as specified in Python's docs. This is equivalent to running Python with the `-u` flag (although it didn't work in this case, hence the environment variable method).