Quotes cog (all of its cmds) set to guild-only
This is because none of the quotes work in DMs. Also the quote cmds relies on other members to write the quote on someone's behalf a lot anyway. In a DM, there's only the bot and user, and while that may still be a valid use case, it wull give a fall-sense of privacy since quotes are private to the servers a particular user is in (with the bot). Therefore, no privacy protection is afforded by DM-ing should the user be in servers with the bot (which will always be the case since there's at least 1 server the user used to DM the bot in the first place*).

*The only time this may possibly not be true is with User-Installed Apps, which we currently don't support.