refactor(quotes): Use a single sender for quote cmd
Let the sender function handle the formatting, and just pass data (but minimize processing on the sender's end, such as passing the name of the member instead of its object). Then, pass the callback func as well which will be called inside the sender.

This is to have one unified place to allow for future conditions such as sending quotes as a text file if its too long (assuming we remove the input char limit on quote-add).