fix: Remove deprecated `datetime.utcnow()` usage
d.py treats naive `datetime` objects as local time, same as how they're treated by many `datetime` methods (including `.timestamp()`) as local times, hence the deprecation of `datetime.utcnow()` because it is naive and the potential bugs arising from it being treated as local time on systems with timezones other than UTC+00:00 (otherwise, local time is effectively same as UTC and so no behind-the-scenes conversion happens which was the behavior before commit 8d47322).

What happens is that behind-the-scenes in the d.py library and the `datetime.timestamp()` method, the naive `datetime.utcnow()` is converted to UTC+00:00 (Unix epoch) in an attempt to compensate for the local time (as it is treated as such), not knowing it is already UTC time. The effect of the conversion is that it becomes UTC time - offset of local timezone. In the case of Asia/Manila where it is UTC+08, the conversion actually results in UTC-08 which is then treated as the UTC time, when it isn't. This is the behavior after commit 8d47322.