Restrict OpenAI cmds to owners only
While it is not possible to obtain statistics on historical costs associated with allowing public usage of OpenAI cmds, well let's just say I want to keep my monthly costs down (in everything). Hence, this commit.

Maybe, I'll revert this once I'm sure I'm capable of paying for the costs with lots of extras on hand.