fix(whisper slash cmd): Owner check incorrectly set
Slash cmds do not respect `commands.is_owner()`. In fact, slash cmds don't really have an equivalent check provided by the library because a owner-only slash cmd is weird. But, we keep it this way because we might still want to use it on other servers as well, and maybe this check is only temporary (see commit 969601f).