build: Don't run GH Action on push
As said in commit 0cc5fe5, this is mainly for installing deps and making sure there are no conflicts. A push either comes from the user or probably Dependabot. In the latter, it's always done through a PR and so the check was already performed, but the GH Action runs again when the Dependabot PR is merged which is unnecessary. In the former, if its a PR then it's fine, but if a direct push to the main branch, the deps have most likely been verified to install correctly locally anyway.