(quotes): Remove blank line between quote add/delete status and quote preview
This results in no line spacing on the web/desktop Discord app, but on mobile, there is still a line spacing. The latter is the reason for this change as the line spacing on mobile Discord app is too great. Visually, no line spacing doesn't look too appealing on web/desktop but its worse to have too big of a line spacing on mobile. So, this is the compromise.