fix(quote & quote-list cmds): Incorrect default member handling
What would happen is that the cmds work as expected when invoked with no member specified in either inside or outside of the listed servers, and also if it was specified in the listed servers. But there existed no code path to handle a specified member outside of the listed servers (due to the wrong invocation of `determine_default_quote_member()` which will return `None` anyway overriding the parameter).