fix(quote cmds): Defer all CRD
This is because sometimes the Postgres DB in neon.tech doesn't go alive in time, resulting to what appears as a fail in the UI.
UI/UX wise, critically needed for quote-delete, but also for quote-add to avoid confusion. Also did it for quote and quote-list so that user doesn't have to retry in case it does fail. The use of `ctx.typing` is also encouraged so that when invoked via traditional cmds, at least there's some indication of progress if the response stalls due to waiting for the DB to come alive. All of this adds overhead, but its neccessary for slash, and provides a better UI/UX thru traditional (instead of a long lag time).